Abstracted the measure function in the radar & passive sensor
