+++ stonesoup/sensor/passive.py
@@ -1,16 +1,14 @@
-from typing import Set, Union
-
 import numpy as np
 
 from ..base import Property
-from ..models.measurement.nonlinear import CartesianToElevationBearing
-from ..sensor.sensor import Sensor
+from ..models.measurement.nonlinear import CartesianToElevationBearing, \
+    CartesianToElevationBearingRange
+from ..sensor.sensor import SimpleSensor
 from ..types.array import CovarianceMatrix
-from ..types.detection import TrueDetection
 from ..types.groundtruth import GroundTruthState
 
 
-class PassiveElevationBearing(Sensor):
+class PassiveElevationBearing(SimpleSensor):
     """A simple passive sensor that generates measurements of targets, using a
     :class:`~.CartesianToElevationBearing` model, relative to its position.
 
@@ -41,28 +39,14 @@ class PassiveElevationBearing(Sensor):
             translation_offset=self.position,
             rotation_offset=self.orientation)
 
-    def measure(self, ground_truths: Set[GroundTruthState], noise: Union[np.ndarray, bool] = True,
-                **kwargs) -> Set[TrueDetection]:
+    def is_detectable(self, state: GroundTruthState) -> bool:
+        measurement_model = CartesianToElevationBearingRange(
+            ndim_state=self.ndim_state,
+            mapping=self.position_mapping,
+            noise_covar=self.noise_covar,
+            translation_offset=self.position,
+            rotation_offset=self.orientation)
 
-        measurement_model = self.measurement_model
-
-        if noise is True and len(ground_truths) > 1:
-            # Pre-fetch noise values when more than one target
-            noise_vectors_iter = iter(measurement_model.rvs(len(ground_truths), **kwargs))
-
-        detections = set()
-        for truth in ground_truths:
-            if noise is True and len(ground_truths) > 1:
-                noise_val = next(noise_vectors_iter)
-            else:
-                noise_val = noise
-
-            measurement_vector = measurement_model.function(truth, noise=noise_val, **kwargs)
-
-            detection = TrueDetection(measurement_vector,
-                                      measurement_model=measurement_model,
-                                      timestamp=truth.timestamp,
-                                      groundtruth_path=truth)
-            detections.add(detection)
-
-        return detections
+        measurement_vector = measurement_model.function(state, noise=False)
+        range_t = measurement_vector[2, 0]  # Elevation(0), Bearing(1), Range(2)
+        return range_t <= self.max_range
